change client project structure and rename folders
