+++ client/src/auth/firebase.js
